move action packages to cluster namespace
