+++ cmd/cl001/ac002/execute/runner.go
@@ -9,7 +9,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/giantswarm/awscnfm/pkg/action"
-	"github.com/giantswarm/awscnfm/pkg/action/ac002"
+	"github.com/giantswarm/awscnfm/pkg/action/cl001/ac002"
 	"github.com/giantswarm/awscnfm/pkg/env"
 )
