+++ cmd/cl001/ac002/explain/runner.go
@@ -11,7 +11,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/giantswarm/awscnfm/pkg/action"
-	"github.com/giantswarm/awscnfm/pkg/action/ac002"
+	"github.com/giantswarm/awscnfm/pkg/action/cl001/ac002"
 )
 
 type runner struct {
